// frontend/src/app/shared/types/ui-states.type.ts
export interface UIStates {
  toggleLanguageMenu: boolean;
  toggleStepper: boolean;
  toggleFilter: boolean;
  toggleLoader: boolean;
  toggleMenu: boolean;
  toggleSelectPersons: boolean;
  toggleSelectDate: boolean;
  isMobile: boolean;
  currentStep: number;
  screenWidth: number;
  screenHeight: number;
  toggleDeparture: boolean;
  toggleDestination: boolean;
  toggleSearchFlight: boolean;
}